Movido currencies individuais para o index
-- Os schemas e as interfaces das currencies individuais (que extendem o CurrencySchema) foram movidos para o index
