--- src/db/models/person/currencies/interface.ts
@@ -1,9 +1,0 @@
-import type { Currency } from './generic'
-
-/**
- * A interface do sub-documento 'currencies' da collection people
- */
-export interface Currencies {
-	bitcoin: Currency
-	nano: Currency
-}
